+++ components/dashboard/index.ts
@@ -1,0 +1,5 @@
+export { MetricCard } from "./MetricCard";
+export { UsageMeter } from "./UsageMeter";
+export { MetricsGrid } from "./MetricsGrid";
+export { ActivityHeatmap } from "./ActivityHeatmap";
+export { TopProjects } from "./TopProjects";